add fixed precision math library (method 1)
When I went to do the implementation for the exp proof of concept, I realized that we
need a method of doing fixed precision math in the style of the crypto domain. The compound
equivalent of this is Exponential.sol which this takes some inspiration from.

During the implementation, I found that unfortunately it is difficult to use traditional
generics and trait bounds for this because in parity's implementation of U256 (for example)
they do not provide interfaces such as `checked_add` that would allow me to use trait bounds
for the generics. Thus, sadly, this approach uses a macro to create a generic fixed precision
decimal type over various implementations of large unsigned values (Uxxx). This only applies
to 512, 256 and 128. 64 and 32 are built-in types and would be incompatible with this macro
which is one of the weaknesses of this approach.

In another PR I will use wrapper types to implement a trait-based generics approach. This
PR is to share this approach and see how it compares with the other trait-based generic
approach.