Fix the build in some circumstances
Right now when running builds in some subdirectories there was an
issue with 'wasm' not found errors. This increases cycle times
because you cannot test in these subdirectories.

The issue was traced back to feature std linking. There are some
interesting sharp corners within the no-std development system and
we just got bit by one of them. The rules are such that when you
write up a crate that you want to be no-std compatible you need to
add a std feature and under that feature you must add all other
crates that have a std feature.

This is a particularly sharp corner because there is no systematic
method for catching this issue at the moment that I am aware of.
Another problem with this issue is that due to conditional compilation
missing a std link in your feature chain results in impossible to
decypher compiler messages.