Substrate Master & Apple Silicon
This patch updates Compound Chain to use the latest Substrate master code (currently linked via local paths, but it will be via git references before this is merged). We upgrade our node template to use the latest template code, which includes a few small changes. Notably, these changes let us use the current nightly branch of Rust, as opposed to the older version from October which was previously referenced in the docs.

Additionally, we are working on minor changes to support Apple Silicon. This is mostly with changes to Substrate to preclude RocksDB compilation but this will be addresed separately. Also, we're currently having issues with `runtime-interfaces` and its interaction with native/wasm.