Substrate Master && Support Apple Silicon (#14)
* Substrate Master & Apple Silicon

This patch updates Compound Chain to use the latest Substrate master code (currently linked via local paths, but it will be via git references before this is merged). We upgrade our node template to use the latest template code, which includes a few small changes. Notably, these changes let us use the current nightly branch of Rust, as opposed to the older version from October which was previously referenced in the docs.

Additionally, we are working on minor changes to support Apple Silicon. This is mostly with changes to Substrate to preclude RocksDB compilation but this will be addresed separately. Also, we're currently having issues with `runtime-interfaces` and its interaction with native/wasm.

* Make runtime benchmarks optional and s/Trait/Config/g

* Switch to using a git-based substrate dep

* Fix remaining minor issues

* patch geoff's code to get runtime_interfaces to build

needed to disable oom and panic handler as it is on master in sp-io
I don't fully understand why that is necessary

* Bump up Cargo.lock

* Make ParityDB the default database, for now

* Fix issue with path deps

* Add back runtime_interfaces HostFunctions

* Try and get CI printing test results

* Fix artifacts

* Push test results to teetotaler

* Fix broken test

Co-authored-by: Wayne Nilsen <wayne@compound.finance>