Magic Extract
This branch is the basis for Magic Extract, which is a function that's meant to be throwaway (literally, we'll delete it the moment it's complete), but with a focus on stressing all functionality of Compound Chain to discuss, in one collective place, all of the decisions that need to be made (e.g. data types for core concepts, interaction between components, error handling libraries, etc). This code is thus meant to be throw-away, but that's not to say we should treat it any less seriously than valid code. This is the place to discuss, argue, biekshed, etc so we don't need to relitigate these issues in future feature branches.