Magic Extract
This branch is the basis for Magic Extract, which is a function that's meant to be throwaway (literally, we'll delete it the moment it's complete), but with a focus on stressing all functionality of Compound Chain to discuss, in one collective place, all of the decisions that need to be made (e.g. data types for core concepts, interaction between components, error handling libraries, etc). This code is thus meant to be throw-away, but that's not to say we should treat it any less seriously than valid code. This is the place to discuss, argue, biekshed, etc so we don't need to relitigate these issues in future feature branches.

Remove Lib

This patch removes `lib.rs` which means the node can't be used as a lib. Alternatively, it also means we don't compile our node twice, which could mean a significant boost in compilation time since this is the highest-level and thus slowest component.

just notices module

stub out actuall processing of notices b/c it opens can of worms about signing

switch to using the account and amount types

Add new types

Add transaction validity check (or remove it really) and fix types

add the submitting of unsigned tx from worker

put notices in Box so it actually runs