Compound Chain DevOps
This patch begins to add deployment operations to Compound Chain such that you can easily deploy a secure Compound Chain node to AWS using Terraform and Ansible. This should be easy for anyone to run to spawn new nodes.

This is a WIP. Please see `README.md` for information on the process. See `ARCHITECTURE.md` for a description of the architecture.