Magic Extract (#26)
* Magic Extract

This branch is the basis for Magic Extract, which is a function that's meant to be throwaway (literally, we'll delete it the moment it's complete), but with a focus on stressing all functionality of Compound Chain to discuss, in one collective place, all of the decisions that need to be made (e.g. data types for core concepts, interaction between components, error handling libraries, etc). This code is thus meant to be throw-away, but that's not to say we should treat it any less seriously than valid code. This is the place to discuss, argue, biekshed, etc so we don't need to relitigate these issues in future feature branches.

* Maybe add a types.json?

* just notices module

* add a notices module

* stub out actuall processing of notices b/c it opens can of worms about signing

* switch to using the account and amount types

* Add new types

* Add transaction validity check (or remove it really) and fix types

* SCALE encoding on Amount type (#34)

* Code snippit for a working scale encoding for Amount type

* SCALE encode/decode for Amount
* this is WIP and not ideal
* will possibly be refactored, mostly sharing here to help unblock max

* improved implementation

* Notice Queue (#31)

add storage for notices

* cargo format on magic-extract branch

* Revert "cargo format on magic-extract branch"

This reverts commit c3f763c2f11289ce1bc65cd1d267d40fbe72eba6.

* WIP: ethereum event processing (#28)

* add cargo fmt check to CI (#37)

* add cargo fmt check to CI

* cargo fmt

* cargo fmt

* add the submitting of unsigned tx from worker (#32)

* notice as enum

cargo fmt

remove app_crypto - it is currently not necessary

doubled up importing on merge

* rebase

* Fetch events in OCWs (#35)

Fetch events with ocw storage lock, version 1

* Notice updates/musings + fix mock/test compilation (#38)

* actually iterate over the notices (#39)

* ECDSA signatures in the style of Ethereum (#41)

* Adds support for signing messages using the correct hash function
* Compatible with the rest of the substrate framework for keys and signing messages
* Currently not really used, to follow up with app_crypto declaration

* Add extrinsic for `process_ethereum_event` (#40)

Add extrinsic for `process_ethereum_event` after fetching `Lock` event, unsigned transaction for now

* Add more config vars (#42)

* Revert "ECDSA signatures in the style of Ethereum (#41)" (#44)

This reverts commit 60f7a03b9bec09722842ab519b0622bbf5a66b16.

* Notice types idea (#43)

Co-authored-by: Coburn Berry <coburn@compound.finance>
Co-authored-by: Max Wolff <maxcwolff@gmail.com>
Co-authored-by: Wayne Nilsen <wayne@compound.finance>
Co-authored-by: Jared Flatow <jflatow@gmail.com>
Co-authored-by: Antonina Norair <antonina.norair@gmail.com>