Separate validator specific configuration runtime interface (#51)
* Separate validator specific configuration runtime interface

The comments in the code generally summarize why this is necessary and what it is. One question that I think is open to debate is "Why use a separate runtime interface?". The answer there is that I did not want any confusion on behalf of the calling side of this code now or in the future. It naturally lends itself to a separate interface with separate functionality. This helps keep the two types of configuration separated in the developer mind. Another question that I think is worth asking is "Why not move eth_rpc_url into the new world?". The answer there is mainly just to avoid being disruptive to those already happily using the code possibly breaking local-dev setups as they exist today.

I was thinking about adding a dot env file but at this point I think that the environment variables that we will be using for configuration should be fairly self documented right here in this module and there should not be a documentation issue.

* move ETH_RPC_URL to an environment variable sooner rather than later

* fix test