Move the initial set of validators to the genesis config
This does have some assertions in it. It follows the pattern from the BABE pallet that also uses assertions. I have tripped these assertions and basically the program crashes when you provide a bad chain spec. This is fairly expected behavior but perhaps not as pretty as it could be. Sadly, I don't think there is anything else we can do with these errors because the macro expects the function provided to `build` to return `()`.

I understand that the shape of the validators may change at some point in the near future but this PR really shows the way to move this configuration into the chain spec file where it may belong. An alternative here is to say "you start with no validators until governance sets the validators". I think this is probably fine but for dev purposes you most likely don't want to have to go through that trouble and this provides a nice escape hatch.