New error handling pattern (#70)
This PR is to demonstrate a new method for handling errors in a consistent way. There are some issues with the way that we handle errors in cash/src/lib.rs right now. The boilerplate of `map_err(|_| ...)` is one issue but the bigger issue is that sometimes errors are logged but often they are not. This can make errors difficult to debug. Using this macro allows us to inject logging for every error that comes up. Using conditional compilation `if cfg!(debug_assertions) {` we can make these logs only happen in debug builds or we can put it behind a fature flag.