Add Integration Tests for Golden Paths
This patch adds integration tests to test the golden path scenario for Compound Chain. The goal is to actually deploy a Cash Token, Starport and run Compound Chain against some local node and test real-life behaviors from JavaScript. The goal is to make sure that all components correctly talk to one another. These tests will obviously be slower to run than unit tests, so we may restrict how and when we run these in CI.