Compound Chain DevOps (#30)
* Compound Chain DevOps

This patch begins to add deployment operations to Compound Chain such that you can easily deploy a secure Compound Chain node to AWS using Terraform and Ansible. This should be easy for anyone to run to spawn new nodes.

This is a WIP. Please see `README.md` for information on the process. See `ARCHITECTURE.md` for a description of the architecture.

* Task part a

* Start to add ansible

* Remove a few manual steps in setup of Ansible

* Start to get a dev server up and running fully

* Big overhaul to Ansible to properly use roles and keys

* Don't purge every time and use one load balancer per protocol

* Update docs

* Updates for 2 terraforms

* Open ephemeral ports since I guess we need them for the tcp connections

* Revise best practices