Add Integration Tests for Golden Paths (#75)
* Add Integration Tests for Golden Paths

This patch adds integration tests to test the golden path scenario for Compound Chain. The goal is to actually deploy a Cash Token, Starport and run Compound Chain against some local node and test real-life behaviors from JavaScript. The goal is to make sure that all components correctly talk to one another. These tests will obviously be slower to run than unit tests, so we may restrict how and when we run these in CI.

* Remove fancy .? syntax

* Add back cargo compilation step

* s/unit/integration/g

* Update README

* Move util functions to own file

* Upgrade test more

* Allow everything for now

* more fixes

* Upgrade to newer version of Substrate

* Fix golden path test for Magic Extract

* add back per extrinsic checking

* Move back to compound-chain#master

* cargo fmt

* Bump down dep as per issue from recent minor package update

* Upgrade mocks for new Substrate version