More setup and progress on the core libs (#78)
* More setup and progress on the core libs

This includes a rethinking/factoring of the types. I also have another `Uint` type to wrap `BigUint` but at the end of the day, biguints kinda suck. You can't construct them as constants and you can't copy them.

So the idea is now to use u128 for amount-like types, and probably BigUint for index-like types (since they grow unboundedly).
I'm going finish the Index type up now and continue wrapping up XXXs, but I wanted to share progress since its getting pretty large and this is going to impact other branches a lot probably.

* Settle on a fixed-width index type as well

While it might be nice to represent indices as unbounded unsigned integers, it doesn't really help if the amount types are not anyway.

* WIP