Scenario Tests
This patch builds scenario tests. The goal is high-level tests for black-box functionality. Instead of building our own language, we opt to use standard JavaScript with a variety of declarative set-up and simplified interfaces to build scenarios that can hopefully cover the majority of behavior of the system before much of the core functionality has been written. For example, a test might look like:

```javascript
{
    name: 'Lock Collateral',
    scenario: async ({ ashley, usdc, chain }) => {
      await ashley.lock(100, usdc);
      expect(await ashley.tokenBalance(usdc)).toEqual(900);
      await chain.waitForEthProcessEvent('cash', 'GoldieLocks'); // Replace with real event
      expect(await ashley.chainBalance(usdc)).toEqual(100);
    }
}
```

The goal is that this test can be clear and obvious in exactly what is occuring, even if much of the actual interaction is hidden behind `ashley.lock()`, etc. We can easily build up into very complex tests from this simple framework.