Add Erc20 Tokens to Integration Tests (#98)
* Add Erc20 Tokens to Integration Tests

This patch adds real Erc20 tokens to our integration tests. This is to mimic the real-world behavior of these tokens in an effort to keep our integration tests as true-to-fact as possible. The integration test framework is now set-up to also be used by the compound-castle repo in development mode, so we'll continue to add features and options to support that repo as well.

* Switch USDC to FiatTokenV1

* Add file that was missed