Add `log!` Macro
This patch adds a `log!("My log {}", 5);` macro that should (hopefully) work correctly in both std and no_std environments. We've historically used a variety of log-styles and the goal here is to consolidate into a single macro that can capture any important log information we need, esp. during the development phase and to debug test-net nodes.