Increment Notice Ids (#127)
* Increment Notice Ids

This patch increments notice ids. Currently, the strategy here is to always increment from (0, 0) to (0, 1), ... and when we hit a new era, that code is responsible for changing the `NextNoticeId` to (1, 0), which will be incremented to (1, 1) then (1, 2) and so on. We could change and evolve this system, but this was an easy approach until other pieces are solidified. Note: none of this is in the spec as it's considered an implementation detail.

* Add a test case for incrementing notice id

* Remove unused function