Simplified Notice Encodings (#128)
* Simplified Notice Encodings

This patch implements the basis for simplified notice encodings, where notices are encoded as Solidity function calls, nearly directly (as notices are specific to a chain to begin with). This only completes Extract (`unlock()`), but it's fairly trivial to convert the rest of the notices.

As a deviation from the current spec, we actually _do_ add to the header of the notice encoding `(eraId, eraIndex, parent)`, as opposed to making the called function (e.g. `unlock`) take those arguments (that is, it's `unlock(address asset, uint amount, address account)` not `unlock(uint eraId, uint eraIndex, bytes32 parent, address asset, uint amount, address account)`. I _believe_ this will make certain procedures (e.g. notice chains) easier to implement, even if it makes this code a little more difficult (since it can't just rely exclusively on Solidity to decode the call parameters). This is something that can be revisited, but if we like it, I will add it back to the spec.

* Start to upgrade and clean up Starport

* Fix test cases for new notice encodings

* Bump Saddle version

* Complete Starport refactor

* Fix bug with reversing eras

* Fix encodings and Rust tests

* Try to bump Node version for CI

* Fix ANTLR and add coverage

* Bump ganache-core package back

* Change case of matchers

* Build coverage report

* Correct directory

* Bump node version for Integratoin tests