Notice Chain
This patch starts to build the notice chain system, such that notices can be accepted even after the authorities have been changed over. Note that we _explicitly_ do not allow you to pass in a notice from a year ago with the notices from back then. Why? Because if the node keys leaked 6 months later then anyone could forge all such notices. The notice chain allows you to pass in that notice with authentication that it was verified _then_.

Currently this is just the basis of the Solidity code for accepting a notice chain, but this patch will, when complete, make sure that notice chains are producted by the runtime and we'll verify passing those in via an integration test.