Handle `lock` and `unlock` CASH (#140)
* Handle `lock` and `unlock` CASH

This patch cleans up the cases of `lock` and `unlock` according to recent changes to the specification. As such, the data to and from Compound Chain use a `principal` instead of `amount`s, which actually simplify the logic. We move amount to principal calcuations into the Cash token, leaving the Starport ideologically pure and unconcerned with that math.

This patch needs significantly more testing.

* Add Extract Cash functionality

* Start to get real Cash extractions to work

* Build Extract Cash Scenario

* Remove only

* Fix Starport tests

* Mind the skips