ADT Events
The majority of the code in the codebase now uses ADTs instead of Generics. The event code was using generics, which led many code paths to be difficult to expand (e.g. get logs returned a `Vec<LockEvent>`). Additionally, it meant that we had storage types like `EthEventQueue` and extrinsics like `process_eth_event`. Switching the code to ADTs, we're able to return a `Vec<ChainEvent>` which can then be for any chain and thus processed downstream as such, as well. Additionally, we start to move code out of `lib.rs` and into `core.rs` as is our standard policy. Finally, we rework event decoding in `compound_crypto` to use the built-in `ethabi` log decoders instead of any custom decoding.

Start to get ADT events passing tests

Fix the rest of the test cases

Switch integration tests to reserved peers to prevent cross-chatter