Add Simple Actor Declarations
This patch just adds a simple process for actors to "state" they are doing something, which is logged to the command line, and subsequently logs when it succeeds or fails. This is in preparation for more complex integration tests where it starts to get _very hard_ to understand the state of user actions when debugging a test ("wait, did Alice's lock go through before Bob tried to borrow?"). This is part of a larger PR that's being broken up to make it easier to merge.