Fix cash index and add logging
The problem here was that the interest index was not being incremented because
the level of precision was too low on CashIndex for a single period of interest.

This PR fixes that. There is probably some math we could do to figure out what
level of precision we need to support the minimum cash interest and block time combination

There really is no reason that we shouldn't have a high number of decimals on
cashIndex because it will be fairly close to 1 for a long time.