Deep Scenarios
This patch is based around making deeper integration tests (scenarios) working for a variety of chain features. E.g. we're starting to test transfering and extraction of Cash, transfer Max, and soon liquidation. Generally these tests are not precise when it comes to interest, but that's not the point of the test, since the precision is lost in the _real flow of time_. Generally acceptance here is what a human could observe when truly interacting with the live system.