Check Failures
This patch is an attempt to improve the failure system overall of Compound Chain. Specifically, we emit events on failure that should make it easy for clients to understand failures. Additionally, as we've solidified around `Reason` as the core error type, we no longer need a macro for error handling. Finally, we fix our event watching system to make it where our clients will be able to get consistent errors without repeats.