Minor Integration Patches
This patch is to help integration tests pass more frequently. There's honestly a lot of work to do as tests get significantly more complex. We now have session changeover, etc, so it's not... not easy to make these consistently pass.

The core hack here is to give events some time when we see them to when we pass them back to the caller s.t. they can be incorporated in a new block. We need a better event polling system or to pair an event polling system with a block polling system to check finality better.