Give Miner Rewards via Inherents
This patch adds a `Miner` storage value that is set by an Inherent. This means that on each block, the block author will set his/her Ethereum address (or potentially any account in the future) that will be used to give block rewards. Note: the address is set _after_ the `on_initialize` code runs, so we actually now store the given rewards for the miner (in re: spreads) and give it to the miner at the start of the next block.