Add Upgradeability to Cash Token
This patch adds upgradeability to the Cash Token. As Cash Token requires initialization, we add a new `initialize` function that can only be called once and is called during initialization of the Cash Token (this is acheived by setting the initial index to 1e18 and rejecting unless the index is 0, its default value). We build the downstream code to use the proper upgradeable cash token.