Fix Lock Scenarios
This patch fixes and builds out just about all of the rest of the lock scenarios. This is one of many patches to come that will try to test full functionality of the system while making sure integration tests _should_ all pass. The only test we've yet skipped here is the "accrue lots of interest" test that will likely require a specially compiled Compound Chain to manually set timestamps, and as such, we leave that test for a later date.