Upgrade via Governance
This patch adds versioning into integration tests, so you can fetch and run any version of Compound Chain, as well as upgrade between versions. In the first iteration of this, we simply test that you can upgrade from v0.01 to the current WASM code. However, we're halted here by the fact that 0.01 does not currently accept the final `set_next_code_via_hash` extrinsic. As such, we'll probably need to add that, cut v0.02 and then test v0.02 to curr.