Allow Excess Signatures
This patch is to allow the Starport to accept excess signatures-- that is, if you provide sigantures that aren't valid (either by duplication or passing in invalid signers, we do not strictly reject the notice. We simply do not count such entries and then tally the total at the end. This is to reduce frictions for users, esp. when the authorities are rotated before a notice is mined but it _might_ still have a sufficient quorum with the new authority set.

Note: should we still fail on duplicates?
Note: this is part of 1.4.1 which will improve Notices generally.