Improve Notice Signing Mechanics
This patch fixes a few items with notice signature collection. First, we make sure to check that signers are current validators before appending signatures to the notice pending set. Secondly, we fix the `recovery` code to fix the check for whether or not we _should_ sign (it was recovering to the wrong address and so the OCW were trying to re-sign every signature every block). Finally, we add full test coverage for this sub-module.