Rename Old Versions to Milestones
This is simply a fix for any code that referenced the old version names into the new ones. We also add a quick script for compiling cargo and ethereum for integration tests since we hop between versions more frequently now and the integration tests always require a re-compile to current code.