Bump to m7, Remove Sudo Pallet, Clean up Chain Spec
This patch, first off, bumps us up to m8 and also removes the `sudo` pallet, which we have not used in a long time and no longer deserves to be part of the overall runtime.

Secondly, we remove the majority of data from the `testnet` chainspec and instead opt to add that data in `chains/build_spec.js` which properly pulls in data like tokens and Starports from the already existing files (i.e. compound-config and `./ethereum/networks/{network}.json`). We make sure these files _are_ checked in to the repo so can reference them quite easily (and since they should rarely change, they are not part of releases but instead just binary relics in the source control itself). This method should be clean and easy to add new nets (e.g. another test-net or mainnet) and should work fairly well with local test-nets if we want to add custom configuration for that.

Next, we remove `LockOld` and `LockCashOld` to match our new deployment.

Additionally, we clean up the `scripts/chains/alice.sh`-type scripts to be unified and start to get them more consistently working. We make sure they use temp directories which should obviate the need for frequent purges, as well.

Finally, we add `rpc.json` to releases since that is now considered a key build artifact per release.