Build Ethereum Contract to test Trx Request from Chain
This patch adds a new contract (and a `TrxRequestBuilder`) which tests executing a trx request from on-chain. The builder is rudimentary, but can be the start of a full-time builder. The contract `Locker` just locks Eth and transfers it to you (the sender)'s account on Gateway. I believe this is currenty failing due to `(Transfer Cash Max 0x...)` failing right now.

Note: the one bug discovered so far is that events from a trx are processed in reverse order. We should ensure we fix that in our new event system.