Liquidate Liquidity Check (#284)
This patch adds a liquidity check to a user to ensure they are underwater before we allow liquidations on the account. There are still other areas to consider (i.e. close factor and collateral quantity), but this is a first start on fixing some bugs in the liquidation function.

Additionally, we add significant test coverage to the `has_liquidity_xxx` functions to ensure they are all individually test in a range of scenarios. We found no serious bugs in that process, but we should consider breaking on some invalid inputs (e.g. for unsupported assets, etc). These are currently left as TODOs.