Handle Unpriced and Mismatched Assets (#285)
* Handle Unpriced and Mismatched Assets

To improve completeness, we add checks for matching asset prices to the assets themselves and ensure prices are set when checking liquidity. We change the behavior from pallet-oracle to return an optional price and use a default zero in the places where it's currently expected. We make no effort here to decide if _that_ is the correct behavior- just to keep the status quo in that regard. A few other call sites may have assumed zero prices that will now receive an error, but that is _probably_ a good thing.

* Remove outdated comments