Factor out `Liquidate.rs` and `BalanceHelpers.rs` with Unit Tests
This patch starts to factor out `liquidate.rs` and `balance_helpers.rs` to their own modules in `internal/`. We significantly increase the test coverage for these functions, finding a few minor cases that could have otherwise panic'd. We thus return `results` here for these functions.

I intend to add unit tests to `liquidate.rs`, but I needed the helper functions to be shared first, which was completed in the current patch.