A ton of minor cleanups and refactorings
Nothing here should change any behavior, but I've started to move the rest of the code out of `core.rs` and into more correspondingly-typed modules. E.g. the asset fetching code has moved to `assets.rs` and the on initialize code to `initialize.rs`. This will open up the door to better unit testing as it's more clear what is and is not being covered by unit tests right now. It's important to merge this quickly as it touches a lot of random lines of code.