Debugging Reorgs (#310)
* Fix Cash RPC, Add Max Ingression Blocks, and Misc. Scenario Fixes

This patch fixes the Cash RPC which was double adding interest to the cash results. We've also added a max ingression blocks since certain scenario tests attempted to lock unpriced assets and thus failed. I would debate after a sufficient number of blocks, we should always just take in a given event from the queue (e.g. 50 or 100 blocks). Finally, we fix a large number of scenarios, including making sure we use `types.json` from earlier versions for upgrade tests. We are still facing a number of issues testing upgrades, and thus those are marked as skipped for now. Finally, we switch back from 1-second blocks to auto-mine for the majority of test cases. This is because they run about 50% faster when auto-mining. To handle the ingression queue code, we just mine 100 blocks whenever we're waiting for Eth events.

* Resolve some XXXs

Co-authored-by: Jared Flatow <jflatow@gmail.com>