Fix m8 Failure (#313)
This patch fixes failures running m8 due to the enum shifting between versions. We add a reserved filler into the enums that can later be set to a different value, e.g. for another chain. We try to fill all of the slots since we would prefer that the numbering of chains stays constant and doesn't shift (e.g. it's #0 for ChainAccounts but #1 for ChainAssets and #5 for ChainBlocks).

Currently we cannot run m7 wasm in the native shell. This is likely due to some incompatability in the host functions versus the previously built wasm code. This new test shows off that behavior (it was broken with a Wasmi Trap). We fix the issue by changing the Dot/Eth order in chains.rs (not sure why this is the fix, but it worked). Note: this test now traps on missing a host function for getting cash data, but this is **not of any concern**. The original error was that blocks could not form, and we have a partial fix here that warrants more follow up. That is: why did this fix work? How can we make it work without changing the enum orders?

Again, the current error in the test is not the issue. The issue is "fixed" as we can mine blocks, but we need to figure out why this fix worked and if we can fix it without the "hack".