Native m8 Shell
This patch adds a scenario where we run a native m8 binary and upgrade _it_ to m9 (as opposed to running an m9 binary with m8 WASM runtime and then do the upgrade there). This can detect anomolies in the node runtime that wouldn't be found by other tests. This test _is passing_ but since it requires a binary build in the release directory that is currently hand-built, it's skipped in CI.