Re-add Gateway addresses, which validators now have
Validators are now required to have a minimum CASH balance in order to set session keys, and therefore validate.
On committing a pipeline change, any account with a balance above the minimum is marked as existing in the Substrate account store.
Any account with a balance below the minumum is purged from the account store.

More test coverage needed.