Upgrade everything we can for a chain reset (#353)
* Re-upgrade Substrate

WIP: CASH benchmarks aren't working for some reason (commented out)
 and test is failing due to another account minimal existence issue

Also had to delete GrandpaRpcHandler, but hopefully not needed now?

* Re-add Gateway addresses, which validators now have

Validators are now required to have a minimum CASH balance in order to set session keys, and therefore validate.
On committing a pipeline change, any account with a balance above the minimum is marked as existing in the Substrate account store.
Any account with a balance below the minumum is purged from the account store.

More test coverage needed.

* Really use min balance path in tests

* Attempt to patch issue blocking ingestion

Only filter the blocks that were supported (not dissented), and only allow support or dissent for a given signer.
The vote is being cast only on the support side, dissent is only derived if the chain receives a disagreement with the first one it received.

* Test m10 patch

* Cargo audit fixes

Also upgrade substrate again in order to bump deps. This is using another substrate branch so as not to break the one others are using.

Couldn't update cranelift-codegen without starting to break stuff in substrate.

* Fix first build from integration

* Start to build code substitutes

* New and improved RPCs (#354)

Co-authored-by: Coburn Berry <coburn@compound.finance>

* Reorg fixes (#349)

* Add two basic parser fuzzers and Fuzzbuzz integration (#320)

* Add two basic parser fuzzers and Fuzzbuzz integration

* Upgrade substrate / fix compilation

Co-authored-by: Everest Munro-Zeisberger <evmunro1@gmail.com>

* Update integration testing readme (#333)

* Fix lock scen

* Build Ethereum Contract to test Trx Request from Chain (#282)

This patch adds a new contract (and a `TrxRequestBuilder`) which tests executing a trx request from on-chain. The builder is rudimentary, but can be the start of a full-time builder. The contract `Locker` just locks Eth and transfers it to you (the sender)'s account on Gateway. I believe this is currenty failing due to `(Transfer Cash Max 0x...)` failing right now.

Note: the one bug discovered so far is that events from a trx are processed in reverse order. We should ensure we fix that in our new event system. (It was)

* Bump rust version in CI and simplify scenario workflow

* Use saddle 0.1.23

* test dissenting (#356)

* Fix some tests

* Mark failing tests for follow-up

* Fix fetch_eth_blocks and formulate reorg issue (#357)

* Fix fetch_eth_blocks and formulate reorg issue

We don't need workers to abstain, since that creates different behaviors depending on where the 'invalid blockchain' is found.

Also, avoid fetching duplicates in formulating the reorg.
Also, really only fetch `slack` blocks, not `slack + 1`.

* Fix dissent and concordance scenario

* Cleanups (#359)

* Fix up scenarios for moon branch

* Temp work on scens

* Start to diagnose all remaining scen issues

* Make the re-org test consistent, if slightly less cool

* Fix TOB #2 (#362)

Co-authored-by: Geoffrey Hayes <hayesgm@gmail.com>
Co-authored-by: Coburn Berry <coburn@compound.finance>
Co-authored-by: Max Wolff <maxcwolff@gmail.com>
Co-authored-by: Everest Munro-Zeisberger <evmunro1@gmail.com>
Co-authored-by: Wayne Nilsen <waynenilsen@gmail.com>
Co-authored-by: Coburn Berry <32463466+coburncoburn@users.noreply.github.com>