Back to StorageLock with increased deadline
Each WASM worker runs in its own thread and does not have shared memory, so the previous fix only worked for native.

Setting to 120s deadline should work most of the time, and when it doesn't it's not the end of the world to ignore a few irrelevant blocks, just inefficient if it happens too often.
The other side effect is that the storage lock persists across restarts, which is the downside of making it too long.